Factor out test cases
